backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
numpy==1.26.2
opencv-python==4.8.1.78  # Fixed version number
opencv-contrib-python==4.8.1.78  # Fixed version number
Pillow==10.1.0
networkx==3.1
matplotlib==3.8.2
scikit-image==0.22.0
aiofiles==23.2.1